+++ pom.xml
@@ -14,6 +14,7 @@
     <rest-assured.version>4.2.0</rest-assured.version>
     <assertj-core.version>3.15.0</assertj-core.version>
     <lombok.version>1.18.10</lombok.version>
+    <testOutputDirectory>${project.build.directory}/test-classes</testOutputDirectory>
   </properties>
 
   <build>
@@ -60,5 +61,15 @@
       <version>2.7.0</version>
     </dependency>
   </dependencies>
+
+  <profiles>
+    <profile>
+      <id>hw9 all tests</id>
+      <properties>
+        <maven-surefire-plugin.suiteXmlFile>${project.build.testOutputDirectory}/testng/hw9.allTests.xml</maven-surefire-plugin.suiteXmlFile>
+      </properties>
+      <activation><activeByDefault>true</activeByDefault></activation>
+    </profile>
+  </profiles>
 
 </project>
