+++ src/main/java/hw9/dto/BoardDTO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class BoardDTO {
-    public static final String NAME = "First test board";
+    public static final String NAME = "First board test";
 
     private String id;
     private String name;
