+++ src/main/java/hw9/service/BoardService.java
@@ -6,27 +6,33 @@ import io.restassured.http.Method;
 import io.restassured.response.Response;
 import java.util.Map;
 
-
-public class BoardService extends CommonService{
+public class BoardService extends CommonService {
 
     public BoardDTO parseBoard(Response response) {
-        return new GsonBuilder().excludeFieldsWithModifiers().create()
-                                .fromJson(response.getBody().asString(), BoardDTO.class);
+
+        return new GsonBuilder()
+            .excludeFieldsWithModifiers()
+            .create()
+            .fromJson(response.getBody().asString(), BoardDTO.class);
     }
 
     public BoardDTO createBoard(Map<String, String> parameters) {
+
         return parseBoard(getWithParams(Method.POST, BOARDS_END_POINT, parameters));
     }
 
     public BoardDTO getBoard(String boardID) {
+
         return parseBoard(getNoParams(Method.GET, BOARDS_END_POINT + boardID));
     }
 
     public BoardDTO updateBoard(Map<String, String> parameters, String boardID) {
+
         return parseBoard(getWithParams(Method.PUT, BOARDS_END_POINT + boardID, parameters));
     }
 
     public Response deleteBoard(String boardID) {
+
         return getNoParams(Method.DELETE, BOARDS_END_POINT + boardID);
     }
 }
