+++ src/test/java/hw9/functional_level/BaseTest.java
@@ -1,8 +1,11 @@
 package hw9.functional_level;
 
 import hw9.assertions.BoardAssertions;
+import hw9.assertions.ListAssertions;
 import hw9.dto.BoardDTO;
+import hw9.dto.ListDTO;
 import hw9.service.BoardService;
+import hw9.service.ListService;
 import io.restassured.response.Response;
 import java.util.HashMap;
 import java.util.Map;
@@ -12,20 +15,25 @@ import org.testng.annotations.BeforeTest;
 public class BaseTest {
 
     public BoardDTO boardDTO;
-    public final BoardService boardService = new BoardService();
+    public BoardService boardService = new BoardService();
     public Map<String, String> parameters = new HashMap<>();
     public BoardAssertions boardAssertions = new BoardAssertions();
 
-    @BeforeTest
-    public void createBoard() {
-        parameters.put("name", BoardDTO.NAME);
-        boardDTO = boardService.createBoard(parameters);
-        boardAssertions.assertBoardName(boardDTO, BoardDTO.NAME);
-    }
+    public ListDTO listDTO;
+    public ListAssertions listAssertions = new ListAssertions();
+    public ListService listService = new ListService();
+    public Map<String, String> bodyParametersForRegistration = new HashMap<>();
 
-    @AfterTest
-    public void deleteBoard() {
-        Response response = boardService.deleteBoard(boardDTO.getId());
-        boardAssertions.confirmBoardDeletedResponse(response);
-    }
+//    @BeforeTest
+//    public void createBoard() {
+//        parameters.put("name", BoardDTO.NAME);
+//        boardDTO = boardService.createBoard(parameters);
+//        boardAssertions.assertBoardName(boardDTO, BoardDTO.NAME);
+//    }
+
+//    @AfterTest
+//    public void deleteBoard() {
+//        Response response = boardService.deleteBoard(boardDTO.getId());
+//        boardAssertions.confirmBoardDeletedResponse(response);
+//    }
 }
