+++ src/test/resources/test.properties
@@ -1,4 +1,2 @@
-#X-RapidAPI-Key=b27426869ad179ee563c7fa16bbb204e
-
 apiKey=
 apiToken=
